Move data_class to util
