+++ ef/config/section.py
@@ -1,6 +1,6 @@
 from collections import namedtuple
 
-from ef.config.data_class import DataClass
+from ef.util.data_class import DataClass
 
 
 class ConfigSection(DataClass):
